requirements.txt
flask==2.3.2
pytesseract==0.3.10
pillow==10.0.0
python-dotenv==1.0.0